vel-pdf-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1
pdfplumber==0.11.4
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1